feat: analytics dashboard, stop repair, circuit breaker, risk controls
Trade Review Dashboard (/analytics):
- New page with filterable trade log (outcome, sector, regime, exit, score)
- Score Analysis tab: win rate + expectancy by score bucket (88-91, 92-94, 95-97, 98-100)
- Hold Time tab: performance by hour bucket (≤24h through 120h+)
- Post-Exit Drift tab: missed winners + exit timing verdict by exit type
- Added /analytics to AppHeader navigation

Bot risk & protection improvements:
- EOD stop fix: eodCancelPartialFills() now preserves all sell-side and stop orders
- Per-cycle stop verification: repairs missing or stale stops every bot cycle
- Morning stop verification: 9:31 AM ET scan places stops for any unprotected positions
- Risk-based position sizing: max $15 risk ÷ stop% = position cap
- Adaptive score thresholds: regime offset applied to minBuyScore (BULL_STRONG -3 → PANIC +10)
- Swing trailing stops: tighter as gains grow (≥20%→2%, ≥10%→3%, <2%→5.5%)
- Daily circuit breaker: halts on 3+ stops OR ≥2% equity loss in a day (DAILY_LOSS_HALT_PCT)

New API endpoints:
- GET /api/performance/score-analysis — score bucket analytics from trades table
- GET /api/performance/hold-time-analysis — hold hour bucket analytics from trades table
- GET /api/ai-trading/portfolio-heat — total open risk in $ and % of equity

Metadata tracking:
- SELL records now carry ai_score, entry_regime, hold_hours, slippage_pct forward
- BUY inserts record entry_regime + slippage_pct derived from notes JSON
- brokerService: added placeStopOrder + cancelOrder to both Alpaca and simulated brokers

New services: premarketGapAlertService, vixSpikeMonitorService

+++ backend/src/config/initDatabase.js
@@ -568,7 +568,7 @@ async function initializeDatabase() {
                 previous_state  TEXT,
                 broker          TEXT,
                 broker_order_id TEXT,
-                quantity        INTEGER,
+                quantity        NUMERIC(14, 8),
                 price           NUMERIC(12, 4),
                 metadata        JSONB       DEFAULT '{}'::jsonb,
                 created_at      TIMESTAMPTZ DEFAULT NOW()
